set hostname to etc

+++ source/vendors/Ralink/RT5350/init/02_etc.sh
@@ -1,4 +1,7 @@
 #!/bin/sh
 
-# linkink profile file
+[ -f /var/run/etc ] && exit
+touch /var/run/etc
+
+# link profile file
 ln -s /etc_ro/profile /etc/profile
